+++ src/CString.h
