+++ src/wpnmod_entity.cpp
