+++ src/wpnmod_entity.h
