+++ src/wpnmod_hooker.cpp
