+++ src/wpnmod_hooker.h
