+++ src/wpnmod_hooks.cpp
