+++ src/wpnmod_hooks.h
