+++ src/wpnmod_memory.cpp
