+++ src/wpnmod_memory.h
