+++ src/wpnmod_natives.cpp
