+++ src/wpnmod_parse.cpp
