+++ src/wpnmod_utils.cpp
