+++ src/wpnmod_utils.h
