+++ src/wpnmod_version.cpp
