+++ src/wpnmod_vtable.cpp
